extract ComponentGroupMapper interface
