+++ components/ledger/ledger-persistence/src/main/kotlin/net/corda/ledger/persistence/consensual/impl/ConsensualComponentGroupMapper.kt
@@ -1,6 +1,6 @@
 package net.corda.ledger.persistence.consensual.impl
 
-import net.corda.ledger.persistence.common.ComponentGroupMapper
+import net.corda.ledger.libs.common.ComponentGroupMapper
 import javax.persistence.Tuple
 
 /**
